Adds conversor and basic style
Change the structure of the html for a basic display.
Adds css.
Adds service to convert from Unix time stamp to Datetime.

+++ WeatherApp/Models/WeatherViewModel.cs
@@ -15,8 +15,8 @@ namespace WeatherApp.Models
         public double TempMin { get; set; }
         public int Pressure { get; set; }
         public int Humidity { get; set; }
-        public int Sunrise { get; set; }
-        public int Sunset { get; set; }
+        public long Sunrise { get; set; }
+        public long Sunset { get; set; }
 
     }
 }
